Responsiveness and Quality of life
Formatted the html, as it was a tad frustrating looking around.

Have added an information modal upon loading the site. Button to hidden and visibile.

Changed colours to CSS variables for easier colour changes.

Added default padding and margin to 0. Made the radio button container(s) into flex and wrapped the radio buttons in div for responsive purposes. Added test meta tags and icon

Styled the input fields to fit the theme more.

1 Issue: The input field in the guide modal, apparently vanished later than the rest - not sure why.